% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipapi-package.R
\docType{package}
\name{ipapi}
\alias{ipapi}
\alias{ipapi-package}
\title{A package to geolocate IPv4/6 addresses and/or domain names using ip-api.com's API}
\description{
A package to geolocate IPv4/6 addresses and/or domain names using ip-api.com's API
}
\author{
Bob Rudis (@hrbrmstr)
}

